refactor(agentic): move wrapper helpers to instrumentation utils
