+++ aws-distro-opentelemetry-node-autoinstrumentation/src/instrumentation/instrumentation-langchain/instrumentation.ts
@@ -15,9 +15,8 @@ import {
   isAgenticInstrumentationOptIn,
   isInstrumentationDisabled,
   detectConflictingInstrumentation,
-  tryWrap,
-  tryUnwrap,
 } from '../../utils';
+import { tryWrap, tryUnwrap } from '../common/instrumentation-utils';
 
 export const INSTRUMENTATION_NAME = '@aws/aws-distro-opentelemetry-instrumentation-langchain';
 export const INSTRUMENTATION_SHORT_NAME = 'aws_langchain';
